Update test to use rewire instead of jest.spy
This is to avoid having to refactor entire files and append `this.` onto
every function that is referenced outside of the exported object.

Due to the way that common JS modules work, there was difficulty using
`jest.spy` when functions were not defined directly within the exports
object and then referenced by another function.

For example, imagine your module has two functions:
```
function foo(){
    bar()
}

function bar() {}

module.exports = {foo, bar}
```

When you try to spy on `bar` with Jest, Jest will replace the
`module.exports.bar` with the mocked function, and will look for it to
be called.  However, `foo` actually calls the `bar` that resides within
the file, rather than the module.exports object property of the same name.

So even though the function gets called, `jest.spy` will say it wasn't,
because it was expecting it to be called as the object's method instead.

Using `jest.spy` will still be useful where it can be used, but rather
than refactor the entire app to export everything and refer to functions
using `this`, we can use rewire for those situations.