Refactor and fix ban and tempban
In my previous attempt, if verifyReasonLength determined that the reason
was too long, send the message but not stop the command execution.

Further, I realized that `msg.content` was the appropriate text to be
measured because that is what gets inserted into the `mod_log` table.

This commit also refactors the tempban SQL statement, which was not
using escaping properly.